Further Email Verification Changes > mail configs added to __init.py__ > isVerified field added to User model... need to migrate then update db before flask run > verification_email.html and verification_email.txt created under templates > login() and register() funcs changes to send_email_verification(). > verify_email() func added to routes.py > send_verification_email() added to email. > email field added to LoginForm- afterwards this can be improved by querying the db for the existing user's email in order to send the verification email, rather than having the user retype their email every time
Current issue: Application-specific password required. Did some investigating and found out that if 2FA is enabled for your Google account this error will happen. Currently trying to create an app specific password right now, but can't seem to do it without setting 2FA on. I believe this is the last issue that plagues this MR...